refactor: 컨벤션 준수 및 폴더 정리
