+++ README.md
@@ -1,5 +1,9 @@
 # 🚇 지하철 노선도 미션
 
+```bash
+
+```
+
 ## 🚀 기능 요구사항
 
 ### 지하철 역 관련 기능
