+++ src/common/checkInput.js
