+++ src/common/clearInput.js
@@ -1,0 +1,4 @@
+export function clearInput() {
+  const $input = document.querySelectorAll("input");
+  $input.forEach((input) => (input.value = ""));
+}
