+++ src/event/line/addLineEvent.js
@@ -4,15 +4,19 @@ import {
   isSameDestination,
   isLineAlreadyExist,
   isTransferStation,
-} from "../../inputCheck.js";
-import { alertMessage } from "../../alertMessage.js";
+} from "../../common/checkInput.js";
+import { alertMessage } from "../../common/alertMessage.js";
+import { clearInput } from "../../common/clearInput.js";
 
 function addLine(line, start, end) {
   let lines = JSON.parse(localStorage.lines);
   let newLine = { name: String(line), sections: [start, end] };
+
   lines.push(newLine);
   localStorage.lines = JSON.stringify(lines);
+
   renderLine();
+  clearInput();
 }
 
 function checkValidLine() {
@@ -20,6 +24,7 @@ function checkValidLine() {
   const start = document.getElementById("line-start-station-selector").value;
   const end = document.getElementById("line-end-station-selector").value;
   const lines = JSON.parse(localStorage.lines);
+
   if (!isSatisfyLength(line)) {
     return alert(alertMessage.SHORT_LENGTH_ERROR);
   } else if (isSameDestination(start, end)) {
@@ -31,6 +36,7 @@ function checkValidLine() {
   } else if (isTransferStation(lines, end)) {
     return alert(`${end}` + alertMessage.TRANSFER_STATION_MESSAGE);
   }
+
   addLine(line, start, end);
 }
