+++ src/event/section/delSectionEvent.js
@@ -1,11 +1,12 @@
 import { renderSection } from "../../render/renderSection.js";
 import { getIndexOfLine } from "./addSectionEvent.js";
-import { isSatisfyMinNumOfStations } from "../../inputCheck.js";
-import { alertMessage } from "../../alertMessage.js";
+import { isSatisfyMinNumOfStations } from "../../common/checkInput.js";
+import { alertMessage } from "../../common/alertMessage.js";
 
 function getIndexOfSection(indexOfLine, section) {
   const lines = JSON.parse(localStorage.lines);
   let indexOfSection;
+
   for (let i = 0; i < lines[indexOfLine].sections.length; i++) {
     if (section === lines[indexOfLine].sections[i]) {
       indexOfSection = i;
@@ -36,11 +37,13 @@ function findDeleteTarget(event, selectedLine) {
   const $sectionTable = document.getElementsByClassName("section-table-row");
   const targetNumber = $target.closest("tr").dataset.number;
   const section = $sectionTable[targetNumber].querySelector("span").innerText;
+
   delSection(selectedLine, section);
 }
 
 export default function delSectionEvent(selectedLine) {
   const $delSectionBtn = document.querySelectorAll(".section-delete-button");
+
   $delSectionBtn.forEach((button) =>
     button.addEventListener("click", (event) => {
       if (confirm(alertMessage.DELETE_CHECK_MESSAGE)) {
