+++ src/render/renderStation.js
@@ -2,11 +2,11 @@ import delStationEvent from "../event/station/delStationEvent.js";
 
 function stationTableTemplate() {
   return `<table class="station-table" border="1">
-    <tr>
-    <th>역 이름</th>
-    <th>설정</th>
-    </tr>
-    </table>`;
+            <tr>
+              <th>역 이름</th>
+              <th>설정</th>
+            </tr>
+          </table>`;
 }
 
 function initStationListContainer() {
@@ -16,13 +16,13 @@ function initStationListContainer() {
 
 function stationListTemplate(station, stationNumber) {
   return `<tr class="station-table-row"data-number=${stationNumber}>
-              <td>
-                <span>${station}</span>
-              </td>
-              <td>
-                <button class="station-delete-button">삭제</button>
-              </td>
-            </tr>`;
+            <td>
+              <span>${station}</span>
+            </td>
+            <td>
+              <button class="station-delete-button">삭제</button>
+            </td>
+          </tr>`;
 }
 
 function initStationList(stations) {
@@ -38,6 +38,7 @@ function initStationList(stations) {
 
 export default function renderStation() {
   initStationListContainer();
+
   const stations = JSON.parse(localStorage.stations);
   if (stations !== null) {
     initStationList(stations);
